Aplicacion/proyecto/backend/requirements.txt
Flask>=2.2.0
psycopg2>=2.9.0
python-dotenv>=0.21.0
